[MIG] hotel_reservation: Migration to 16.0
