+++ hotel_reservation/README.rst
@@ -7,7 +7,7 @@ Hotel Reservation Management
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:586817e120c9e7c6b2682503a32a49b256e0b8f0179fe671cbbd3cd3eab11233
+   !! source digest: sha256:46151aa3789bb213451e52ad3b2720039958ae55d9c2b319394bf4936b462e63
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -16,14 +16,14 @@ Hotel Reservation Management
 .. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fvertical--hotel-lightgray.png?logo=github
-    :target: https://github.com/OCA/vertical-hotel/tree/15.0/hotel_reservation
-    :alt: OCA/vertical-hotel
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fpurchase--workflow-lightgray.png?logo=github
+    :target: https://github.com/OCA/purchase-workflow/tree/16.0/hotel_reservation
+    :alt: OCA/purchase-workflow
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/vertical-hotel-15-0/vertical-hotel-15-0-hotel_reservation
+    :target: https://translation.odoo-community.org/projects/purchase-workflow-16-0/purchase-workflow-16-0-hotel_reservation
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
-    :target: https://runboat.odoo-community.org/builds?repo=OCA/vertical-hotel&target_branch=15.0
+    :target: https://runboat.odoo-community.org/builds?repo=OCA/purchase-workflow&target_branch=16.0
     :alt: Try me on Runboat
 
 |badge1| |badge2| |badge3| |badge4| |badge5|
@@ -73,10 +73,10 @@ You can manage:
 Bug Tracker
 ===========
 
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/vertical-hotel/issues>`_.
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/purchase-workflow/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/vertical-hotel/issues/new?body=module:%20hotel_reservation%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/purchase-workflow/issues/new?body=module:%20hotel_reservation%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -96,6 +96,10 @@ Contributors
 * Serpent Consulting Services Pvt. Ltd.
 * Odoo S.A.
 
+* `Trey <https://www.trey.es>`__:
+
+  * Vicent Cubells
+
 Maintainers
 ~~~~~~~~~~~
 
@@ -109,6 +113,14 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/vertical-hotel <https://github.com/OCA/vertical-hotel/tree/15.0/hotel_reservation>`_ project on GitHub.
+.. |maintainer-cubells| image:: https://github.com/cubells.png?size=40px
+    :target: https://github.com/cubells
+    :alt: cubells
+
+Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
+
+|maintainer-cubells| 
+
+This module is part of the `OCA/purchase-workflow <https://github.com/OCA/purchase-workflow/tree/16.0/hotel_reservation>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
