+++ hotel_reservation/data/hotel_reservation_data.xml
@@ -1,5 +1,4 @@
 <odoo noupdate="1">
-    <!-- Demo data for hotel reservation -->
     <record id="hotel_reservation_0" model="hotel.reservation">
         <field
             eval="(DateTime.today()).strftime('%Y-%m-%d 10:00:00')"
