+++ hotel_reservation/data/hotel_scheduler.xml
@@ -1,5 +1,4 @@
 <odoo>
-    <!-- Scheduler For Hotel Room -->
     <record model="ir.cron" id="room_analytic_cron">
         <field name="name">Room Analysis</field>
         <field name="model_id" ref="model_hotel_room" />
@@ -10,7 +9,6 @@
         <field name="numbercall">-1</field>
         <field name="doall" eval="False" />
     </record>
-    <!-- Scheduler For To Inform Guest About Reservation Before 24 Hours -->
     <record model="ir.cron" id="Guest_reservation_reminder_24hrs">
         <field name="name">Hotel Reservation: Booking Reminder before 24 Hours
         </field>
