+++ hotel_reservation/models/__init__.py
@@ -1,6 +1,10 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import hotel_folio
-from . import hotel_quick_reservation
+from . import hotel_folio_line
 from . import hotel_reservation
+from . import hotel_reservation_line
 from . import hotel_room
+from . import hotel_room_reservation_line
+from . import quick_room_reservation
+from . import room_reservation_summary
