+++ hotel_reservation/models/hotel_reservation_line.py
@@ -1,0 +1,97 @@
+# Copyright (C) 2022-TODAY Serpent Consulting Services Pvt. Ltd. (<http://www.serpentcs.com>).
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+
+from odoo import _, api, fields, models
+from odoo.exceptions import ValidationError
+
+
+class HotelReservationLine(models.Model):
+    _name = "hotel.reservation.line"
+    _description = "Reservation Line"
+
+    name = fields.Char()
+    line_id = fields.Many2one("hotel.reservation")
+    reserve = fields.Many2many(
+        "hotel.room",
+        "hotel_reservation_line_room_rel",
+        "hotel_reservation_line_id",
+        "room_id",
+        domain="[('isroom','=',True),\
+                               ('categ_id','=',categ_id)]",
+    )
+    categ_id = fields.Many2one("hotel.room.type", "Room Type")
+
+    @api.onchange("categ_id")
+    def on_change_categ(self):
+        """
+        When you change categ_id it check checkin and checkout are
+        filled or not if not then raise warning
+        -----------------------------------------------------------
+        @param self: object pointer
+        """
+        if not self.line_id.checkin:
+            raise ValidationError(
+                _(
+                    """Before choosing a room,\n You have to """
+                    """select a Check in date or a Check out """
+                    """ date in the reservation form."""
+                )
+            )
+        hotel_room_ids = self.env["hotel.room"].search(
+            [("room_categ_id", "=", self.categ_id.id)]
+        )
+        room_ids = []
+        for room in hotel_room_ids:
+            assigned = False
+            for line in room.room_reservation_line_ids.filtered(
+                lambda l: l.status != "cancel"
+            ):
+                if self.line_id.checkin and line.check_in and self.line_id.checkout:
+                    if (
+                        self.line_id.checkin <= line.check_in <= self.line_id.checkout
+                    ) or (
+                        self.line_id.checkin <= line.check_out <= self.line_id.checkout
+                    ):
+                        assigned = True
+                    elif (line.check_in <= self.line_id.checkin <= line.check_out) or (
+                        line.check_in <= self.line_id.checkout <= line.check_out
+                    ):
+                        assigned = True
+            for rm_line in room.room_line_ids.filtered(lambda l: l.status != "cancel"):
+                if self.line_id.checkin and rm_line.check_in and self.line_id.checkout:
+                    if (
+                        self.line_id.checkin
+                        <= rm_line.check_in
+                        <= self.line_id.checkout
+                    ) or (
+                        self.line_id.checkin
+                        <= rm_line.check_out
+                        <= self.line_id.checkout
+                    ):
+                        assigned = True
+                    elif (
+                        rm_line.check_in <= self.line_id.checkin <= rm_line.check_out
+                    ) or (
+                        rm_line.check_in <= self.line_id.checkout <= rm_line.check_out
+                    ):
+                        assigned = True
+            if not assigned:
+                room_ids.append(room.id)
+        domain = {"reserve": [("id", "in", room_ids)]}
+        return {"domain": domain}
+
+    def unlink(self):
+        ReservationLine = self.env["hotel.room.reservation.line"]
+        for reserv_rec in self:
+            for rec in reserv_rec.reserve:
+                myobj = ReservationLine.search(
+                    [
+                        ("room_id", "=", rec.id),
+                        ("reservation_id", "=", reserv_rec.line_id.id),
+                    ]
+                )
+                if myobj:
+                    rec.write({"isroom": True, "status": "available"})
+                    myobj.unlink()
+        return super().unlink()
