+++ hotel_reservation/readme/CONTRIBUTORS.rst
@@ -1,3 +1,7 @@
 * Odoo Community Association (OCA)
 * Serpent Consulting Services Pvt. Ltd.
 * Odoo S.A.
+
+* `Trey <https://www.trey.es>`__:
+
+  * Vicent Cubells
