+++ hotel_reservation/report/checkout_report_template.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <!--Template for Hotel Reservation -->
     <template id="report_checkout_qweb">
         <t t-call="web.html_container">
             <t t-call="web.external_layout">
