+++ hotel_reservation/report/report_view.xml
@@ -1,15 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <!--Report for hotel check-in details -->
-  <!--   <report
-        id="hotel_checkin_details"
-        string="CheckIn Detail"
-        model="hotel.reservation"
-        report_type="qweb-pdf"
-        file="hotel_reservation.report_checkin_qweb"
-        name="hotel_reservation.report_checkin_qweb"
-        menu="True"
-    /> -->
     <record id="hotel_checkin_details" model="ir.actions.report">
         <field name="name">CheckIn Detail</field>
         <field name="model">hotel.reservation</field>
@@ -17,16 +7,6 @@
         <field name="report_file">hotel_reservation.report_checkin_qweb</field>
         <field name="report_name">hotel_reservation.report_checkin_qweb</field>
     </record>
-    <!--Report for hotel room reservation details -->
-    <!-- <report
-        id="hotel_roomres_details"
-        string="Room Reservation Detail"
-        model="hotel.reservation"
-        report_type="qweb-pdf"
-        file="hotel_reservation.report_room_reservation_qweb"
-        name="hotel_reservation.report_room_reservation_qweb"
-        menu="True"
-    /> -->
     <record id="hotel_roomres_details" model="ir.actions.report">
         <field name="name">Room Reservation Detail</field>
         <field name="model">hotel.reservation</field>
@@ -34,16 +14,6 @@
         <field name="report_file">hotel_reservation.report_room_reservation_qweb</field>
         <field name="report_name">hotel_reservation.report_room_reservation_qweb</field>
     </record>
-    <!--Report for hotel check-out details -->
-  <!--   <report
-        id="hotel_checkout_details"
-        string="CheckOut Detail"
-        model="hotel.reservation"
-        report_type="qweb-pdf"
-        file="hotel_reservation.report_checkout_qweb"
-        name="hotel_reservation.report_checkout_qweb"
-        menu="True"
-    /> -->
     <record id="hotel_checkout_details" model="ir.actions.report">
         <field name="name">CheckOut Detail</field>
         <field name="model">hotel.reservation</field>
@@ -51,16 +21,6 @@
         <field name="report_file">hotel_reservation.report_checkout_qweb</field>
         <field name="report_name">hotel_reservation.report_checkout_qweb</field>
     </record>
-    <!--Report for hotel Maximum used room details -->
-    <!-- <report
-        id="hotel_maxroom_details"
-        string="MaxRoom Detail"
-        model="hotel.reservation"
-        report_type="qweb-pdf"
-        file="hotel_reservation.report_maxroom_qweb"
-        name="hotel_reservation.report_maxroom_qweb"
-        menu="True"
-    /> -->
     <record id="hotel_maxroom_details" model="ir.actions.report">
         <field name="name">MaxRoom Detail</field>
         <field name="model">hotel.reservation</field>
