+++ hotel_reservation/security/ir.model.access.csv
@@ -6,8 +6,6 @@ access_hotel_reservation_manager,hotel_reservation.manager,model_hotel_reservati
 access_hotel_reservation_line_manager,hotel.reservation.line.manager,model_hotel_reservation_line,hotel.group_hotel_manager,1,1,1,1
 access_hotel_room_reservation_line_manager,hotel_room_reservation.line.manager,model_hotel_room_reservation_line,hotel.group_hotel_manager,1,1,1,1
 access_room_reservation_summary_manager,hotel_room_reservation_summary.manager,model_room_reservation_summary,hotel.group_hotel_manager,1,1,1,1
-
-
 access_hotel_reservation_wizard,access_hotel_reservation_wizard,model_hotel_reservation_wizard,base.group_user,1,1,1,1
 access_wizard_make_folio,access_wizard_make_folio,model_wizard_make_folio,base.group_user,1,1,1,1
 access_quick_room_reservation,access_quick_room_reservation,model_quick_room_reservation,base.group_user,1,1,1,1
