+++ hotel_reservation/static/src/css/room_summary.css
@@ -28,39 +28,3 @@
     vertical-align: middle;
     cursor: pointer;
 }
-
-/*$color_1: white;
-$background-color_1: red;
-$background-color_2: blue;
-$background-color_3: green;
-
-.table_header {
-    text-align: center;
-    min-width: 100px;
-    vertical-align: middle;
-}
-.table_reserved {
-    background-color: $background-color_1;
-    color: $color_1;
-    text-align: center;
-    min-width: 100px;
-    vertical-align: middle;
-    cursor: pointer;
-}
-.table_reserved1 {
-    background-color: $background-color_2;
-    color: $color_1;
-    text-align: center;
-    min-width: 100px;
-    vertical-align: middle;
-    cursor: pointer;
-}
-.table_free {
-    background-color: $background-color_3;
-    color: $color_1;
-    text-align: center;
-    min-width: 100px;
-    vertical-align: middle;
-    cursor: pointer;
-}
-*/
