+++ hotel_reservation/static/src/xml/hotel_room_summary.xml
@@ -1,27 +1,36 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <templates xml:space="preserve">
-    <t t-name="RoomSummary">
+    <t t-name="RoomSummary" owl="1">
         <div style="overflow:auto;">
-            <table border="1">
-                <t t-foreach="widget.get('summary_header')" t-as="header_detail">
+            <table border="1" t-ref="textarea">
+                <t
+                    t-foreach="state.summary_header"
+                    t-as="header_detail"
+                    t-key="header_detail_index"
+                >
                     <tr>
-                        <t t-foreach="header_detail.header" t-as="date">
+                        <t
+                            t-foreach="header_detail.header"
+                            t-as="date"
+                            t-key="date_index"
+                        >
                             <th class="table_header" style="text-align:center;">
                                 <t t-esc="date" />
                             </th>
                         </t>
                     </tr>
                 </t>
-                <t t-foreach="widget.get('room_summary')" t-as="detail">
+                <t t-foreach="state.room_summary" t-as="detail" t-key="detail_index">
                     <tr>
                         <td style="text-align:center;">
                             <t t-esc="detail.name" />
                         </td>
-                        <t t-foreach="detail.value" t-as="status">
+                        <t t-foreach="detail.value" t-as="status" t-key="status_index">
                             <t t-if="status.state == 'Free'">
                                 <td
                                     class="table_free"
                                     t-att-data="status.room_id"
+                                    t-on-click="(ev) => this.load_form(status.room_id,status.date, ev)"
                                     t-att-date="status.date"
                                     style="text-align:center;"
                                 >
