+++ hotel_reservation/wizards/__init__.py
@@ -1,3 +1,4 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import hotel_reservation_wizard
+from . import wizard_make_folio
