+++ hotel_reservation/wizards/hotel_reservation_wizard.py
@@ -8,8 +8,14 @@ class HotelReservationWizard(models.TransientModel):
     _name = "hotel.reservation.wizard"
     _description = "Allow to generate a reservation"
 
-    date_start = fields.Datetime("Start Date", required=True)
-    date_end = fields.Datetime("End Date", required=True)
+    date_start = fields.Datetime(
+        string="Start Date",
+        required=True,
+    )
+    date_end = fields.Datetime(
+        string="End Date",
+        required=True,
+    )
 
     def report_reservation_detail(self):
         data = {
@@ -50,28 +56,3 @@ class HotelReservationWizard(models.TransientModel):
         return self.env.ref("hotel_reservation.hotel_maxroom_details").report_action(
             self, data=data
         )
-
-
-class MakeFolioWizard(models.TransientModel):
-    _name = "wizard.make.folio"
-    _description = "Allow to generate the folio"
-
-    grouped = fields.Boolean("Group the Folios")
-
-    def make_folios(self):
-        reservation_obj = self.env["hotel.reservation"]
-        newinv = [
-            order.id
-            for order in reservation_obj.browse(
-                self.env.context.get("active_ids", [])
-            ).mapped("folio_id")
-        ]
-        return {
-            "domain": "[('id','in', [" + ",".join(map(str, newinv)) + "])]",
-            "name": "Folios",
-            "view_type": "form",
-            "view_mode": "tree,form",
-            "res_model": "hotel.folio",
-            "view_id": False,
-            "type": "ir.actions.act_window",
-        }
