+++ hotel_reservation/wizards/hotel_reservation_wizard_views.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <!--Form view for hotel reservation wizard -->
     <record id="hotel_reservation_wizard_form_view" model="ir.ui.view">
-        <field name="name">hotel.reservation.wizard.form</field>
         <field name="model">hotel.reservation.wizard</field>
         <field name="arch" type="xml">
             <form string="Reservation List">
@@ -49,7 +47,6 @@
             </form>
         </field>
     </record>
-    <!--Action for hotel reservation wizard -->
     <record id="action_hotel_reservation_wizard" model="ir.actions.act_window">
         <field name="name">Hotel Reservation Report</field>
         <field name="res_model">hotel.reservation.wizard</field>
@@ -57,46 +54,6 @@
         <field name="view_mode">form</field>
         <field name="target">new</field>
     </record>
-    <!--Form view for wizard make folio -->
-    <record id="make_folio_wizard_form_view" model="ir.ui.view">
-        <field name="name">make.folio.wizard.form</field>
-        <field name="model">wizard.make.folio</field>
-        <field name="arch" type="xml">
-            <form string="Create Folio" version='8.0'>
-                <separator
-                    colspan="4"
-                    string="Do you really want to create the Folio ?"
-                />
-                <!-- <field name="grouped" /> -->
-                <newline />
-                <footer>
-                    <button
-                        name="make_folios"
-                        string="Create Folio"
-                        type="object"
-                        icon="fa-files-o"
-                        class="oe_highlight"
-                    />
-                    <button
-                        string="Cancel"
-                        icon="fa-close"
-                        class="btn btn-primary"
-                        special="cancel"
-                    />
-                </footer>
-            </form>
-        </field>
-    </record>
-
-    <!--Action for hotel reservation wizard -->
-    <record id="act_make_folio" model="ir.actions.act_window">
-        <field name="name">Make Folios</field>
-        <field name="res_model">wizard.make.folio</field>
-        <field name="binding_view_types">form</field>
-        <field name="view_mode">form</field>
-        <field name="target">new</field>
-    </record>
-
     <menuitem
         id="wizard_res_menu"
         name="Hotel Reservation Report"
