+++ hotel_reservation/wizards/wizard_make_folio_views.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="make_folio_wizard_form_view" model="ir.ui.view">
+        <field name="name">make.folio.wizard.form</field>
+        <field name="model">wizard.make.folio</field>
+        <field name="arch" type="xml">
+            <form string="Create Folio" version='8.0'>
+                <separator
+                    colspan="4"
+                    string="Do you really want to create the Folio ?"
+                />
+                <!-- <field name="grouped" /> -->
+                <newline />
+                <footer>
+                    <button
+                        name="make_folios"
+                        string="Create Folio"
+                        type="object"
+                        icon="fa-files-o"
+                        class="oe_highlight"
+                    />
+                    <button
+                        string="Cancel"
+                        icon="fa-close"
+                        class="btn btn-primary"
+                        special="cancel"
+                    />
+                </footer>
+            </form>
+        </field>
+    </record>
+    <record id="act_make_folio" model="ir.actions.act_window">
+        <field name="name">Make Folios</field>
+        <field name="res_model">wizard.make.folio</field>
+        <field name="binding_view_types">form</field>
+        <field name="view_mode">form</field>
+        <field name="target">new</field>
+    </record>
+</odoo>
